further tweaking/unifying of edge routing with ports - backtracking on port definition - I realized that the current port is already defined in the edge - what's needed is really only (optional) port constraints that tell layouts where to connect those edges when being re-routed
git-svn-id: https://svn.code.sf.net/p/genj/code/trunk@5406 fbd980a2-0dff-11df-b4b7-0b05b74d0985

+++ graphj/src/shell/gj/shell/model/EditableEdge.java
@@ -72,8 +72,8 @@ public class EditableEdge implements Edge {
   
   private Path makeShape() {
     return LayoutHelper.getPath(
-        start.getPosition(), start.getShape(), new Point2D.Double(0,0),
-        end.getPosition(), end.getShape(), new Point2D.Double(0,0)
+        start.getPosition(), start.getShape(), new Point2D.Double(),
+        end.getPosition(), end.getShape(), new Point2D.Double()
     );   
   }
